Privatize route_stats: drop from Eventstream, move to an internal helper
route_stats existed only to back the transition graph's route badge --
unlike the widget's other headless twins (transition_graph_data), it had
no standalone analytical purpose worth public API surface, and the graph
already exposes richer per-route detail than the raw dict it returned.

Moved tools/route_stats.py's RouteStats dataclass to a plain function in
utils/route_stats.py (no Eventstream method wraps it, so no `_tracked`
telemetry either -- that only applied to the public headless call).
widgets/transition_graph.py's route_stats compute tool now calls it
directly. Documented this as the one exception to compute_tools' "tool
name mirrors an Eventstream method" convention in widgets/_base.py.

Docs and the 5.1.0 CHANGELOG entry no longer advertise it as a headless
Eventstream method.
